+++ src/algebra/Matrix.cpp
@@ -64,7 +64,7 @@ Matrix Matrix::operator*(const Matrix& other) {
     return newMatrix;
 }
 
-double& Matrix::getAt(int x, int y) {
+float& Matrix::getAt(int x, int y) {
     return matrix[x][y];
 }
 
@@ -97,10 +97,10 @@ void Matrix::resize(int columns, int rows) {
     }
 }
 
-std::vector<double>& Matrix::operator[](int pos) {
+std::vector<float>& Matrix::operator[](int pos) {
     return matrix.at(pos);
 }
-std::vector<double> Matrix::operator[](int pos) const {
+std::vector<float> Matrix::operator[](int pos) const {
     return matrix[pos];
 }
 std::vector<double> Matrix::getRow(int row) const {
